api/app/requirements.txt
Flask
pymysql
apiflask
pandas
python-dotenv
sqlalchemy
sqlalchemy_utils
watchdog
colorama